treating guid and other params separately

+++ setup.sh
@@ -25,37 +25,30 @@ if [ ! -f "$PRIVATE_DIR/guid" ]; then
 
 	echo "$GUID" > "$PRIVATE_DIR/guid"
 	echo "$TOKEN" > "$PRIVATE_DIR/token"
-
-	HOSTNAME="https://api.rfcx.org"
-	echo "$HOSTNAME" > "$PRIVATE_DIR/hostname"
-
-	FILENAME_TIMESTAMP_FORMAT="SCW1840_%Y%Y%m%d_%H%M%S"
-	echo "$FILENAME_TIMESTAMP_FORMAT" > "$PRIVATE_DIR/filename_timestamp_format"
-
-	echo " - Guardian: $GUID"
-	echo " - Token: [secret]"
-	echo " - RFCx API: $HOSTNAME"
-	echo " - Filename Timestamp Format: '$FILENAME_TIMESTAMP_FORMAT'"
-	echo " - "
-
 else
-
 	GUID=`cat "$PRIVATE_DIR/guid";`;
 	TOKEN=`cat "$PRIVATE_DIR/token";`;
-	HOSTNAME=`cat "$PRIVATE_DIR/hostname";`;
-	FILENAME_TIMESTAMP_FORMAT=`cat "$PRIVATE_DIR/filename_timestamp_format";`;
-
-	echo " - This Guardian has previously been setup"
-	echo " - "
-	echo " - Guardian: $GUID"
-	echo " - Token: [secret]"
-	echo " - RFCx API: $HOSTNAME"
-	echo " - Filename Timestamp Format: '$FILENAME_TIMESTAMP_FORMAT'"
-	echo " - "
-
 fi
 
+if [ ! -f "$PRIVATE_DIR/hostname" ]; then 
+	HOSTNAME="https://api.rfcx.org"
+	echo "$HOSTNAME" > "$PRIVATE_DIR/hostname"
+else
+	HOSTNAME=`cat "$PRIVATE_DIR/hostname";`;
+fi
 
+if [ ! -f "$PRIVATE_DIR/filename_timestamp_format" ]; then 
+	FILENAME_TIMESTAMP_FORMAT="SCW1840_%Y%Y%m%d_%H%M%S"
+	echo "$FILENAME_TIMESTAMP_FORMAT" > "$PRIVATE_DIR/filename_timestamp_format"
+else
+	FILENAME_TIMESTAMP_FORMAT=`cat "$PRIVATE_DIR/filename_timestamp_format";`;
+fi
+
+echo " - Guardian: $GUID"
+echo " - Token: [secret]"
+echo " - RFCx API: $HOSTNAME"
+echo " - Filename Timestamp Format: '$FILENAME_TIMESTAMP_FORMAT'"
+echo " - "
 
 
 if [ ! -f "$PRIVATE_DIR/registered" ]; then
